[fix] Separated BaseTechObjectManager to BaseTechObjectLoader and BaseTechObjectManager
